separated physics in 2d and 3d domains
