+++ include/NovelRT/Physics/2D/JointDefinition2D.h
@@ -1,14 +1,14 @@
 // Copyright © Matt Jones and Contributors. Licensed under the MIT Licence (MIT). See LICENCE.md in the repository root
 // for more information.
 
-#ifndef NOVELRT_PHYSICS_JOINTDEFINITION2D_H
-#define NOVELRT_PHYSICS_JOINTDEFINITION2D_H
+#ifndef NOVELRT_PHYSICS_PHYSICS2D_JOINTDEFINITION2D_H
+#define NOVELRT_PHYSICS_PHYSICS2D_JOINTDEFINITION2D_H
 
 #ifndef NOVELRT_PHYSICS_H
 #error NovelRT does not support including types explicitly by default. Please include Physics.h instead for the Physics namespace subset.
 #endif
 
-namespace NovelRT::Physics
+namespace NovelRT::Physics::Physics2D
 {
     struct JointDefinition2D
     {
@@ -23,4 +23,4 @@ namespace NovelRT::Physics
     };
 }
 
-#endif // !NOVELRT_PHYSICS_JOINTDEFINITION2D_H
+#endif // !NOVELRT_PHYSICS_PHYSICS2D_JOINTDEFINITION2D_H
