+++ include/NovelRT/Physics/2D/RigidBody2D.h
@@ -1,10 +1,10 @@
 // Copyright © Matt Jones and Contributors. Licensed under the MIT Licence (MIT). See LICENCE.md in the repository root
 // for more information.
 
-#ifndef NOVELRT_PHYSICS_RIGIDBODY2D_H
-#define NOVELRT_PHYSICS_RIGIDBODY2D_H
+#ifndef NOVELRT_PHYSICS_PHYSICS2D_RIGIDBODY2D_H
+#define NOVELRT_PHYSICS_PHYSICS2D_RIGIDBODY2D_H
 
-namespace NovelRT::Physics
+namespace NovelRT::Physics::Physics2D
 {
     class RigidBody2D
     {
