+++ include/NovelRT/Physics/Physics.h
@@ -17,47 +17,11 @@ namespace NovelRT::Physics
     enum class BodyType;
     enum class JointType;
     class PhysicsProvider;
-    class PhysicsProvider2D;
-    class PhysicsWorld2D;
-    class RigidBody2D;
-    struct BodyDefinition2D;
-    class Joint2D;
-    class DistanceJoint2D;
-    class GearJoint2D;
-    class FixedDistanceJoint2D;
-    class HingeJoint2D;
-    class SlidingJoint2D;
-    class SpringJoint2D;
-    struct JointDefinition2D;
-    struct DistanceJointDefinition2D;
-    struct GearJointDefinition2D;
-    struct FixedDistanceJointDefinition2D;
-    struct HingeJointDefinition2D;
-    struct SlidingJointDefinition2D;
-    struct SpringJointDefinition2D;
 }
 
 #include "BodyType.h"
 #include "JointType.h"
 #include "PhysicsProvider.h"
-#include "PhysicsProvider2D.h"
-#include "PhysicsWorld2D.h"
-#include "RigidBody2D.h"
-#include "BodyDefinition2D.h"
-#include "Joint2D.h"
-#include "DistanceJoint2D.h"
-#include "FixedDistanceJoint2D.h"
-#include "GearJoint2D.h"
-#include "HingeJoint2D.h"
-#include "SlidingJoint2D.h"
-#include "SpringJoint2D.h"
-#include "JointDefinition2D.h"
-#include "DistanceJointDefinition2D.h"
-#include "FixedDistanceJointDefinition2D.h"
-#include "HingeJointDefinition2D.h"
-#include "GearJointDefinition2D.h"
-#include "SlidingJointDefinition2D.h"
-#include "SpringJointDefinition2D.h"
-
+#include "2D/Physics.2D.h"
 
 #endif // !NOVELRT_PHYSICS_H
